firestore.providerの組み込み完了。 third画面(firestore)でリロード状態になるのでそこの原因を見つける

+++ lib/second_page/weight_data.dart
@@ -1,9 +1,9 @@
 ///日付と体重を持つクラスを作成
 class WeightData {
   DateTime? date;
-  int? weight;
+  double? weight;
 
-  WeightData(DateTime date, int weight) {
+  WeightData(DateTime date, double weight) {
     this.date = date;
     this.weight = weight;
   }
